Factoring out instrumentation from TruChain.
